updated files in order to support -Werror for ICC
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2798 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/grid/common/Stretching.hpp
@@ -220,7 +220,7 @@ namespace sg {
          * @param dim dimension index of the type array
          *
          */
-        Stretching1D getStretching1D(int dim);
+        Stretching1D getStretching1D(size_t dim);
 
         /*
          * Prints the lookup table generated. Used for debugging
